lnwallet: add MusigPartialSig wrapper over musig2.PartialSignature
In this commit, we add a small wrapper struct that implements the input.Signature interface. This'll let us drop this in anywhere the ECDSA signatures are currently accepted.

When we serialize the signatures, as we want to pack it in 64 bytes, we use just the x-coordinate of the R point. Note that this means we aren't able to fully reconstruct the R point from just that value (even vs odd). This value is used as input to the challenge hash for verification, and we can recompute R (the full point), given: the combined nonces, the combined key, and the message.
